% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_diff_stats.R
\name{compute_diff_list}
\alias{compute_diff_list}
\title{Compute differentiation statistics \eqn{F_{ST}}, \eqn{G'_{ST}}, and \eqn{D}
for a list of data frames of allele frequencies or subclonal fractions.
The differentiation statistics are computed across all available subpopulations.}
\usage{
compute_diff_list(data_subclonal, col_indices)
}
\arguments{
\item{data_subclonal}{A list of data frames containing allele frequencies or subclonal fractions.}

\item{col_indices}{Indices of columns corresponding to frequencies in subpopulations.}
}
\value{
A named list containing frequency of the most frequent allele \eqn{M}, \eqn{F_{ST}}, \eqn{G'_{ST}}, and \eqn{D}
computed for each locus in each dataset.
}
\description{
Compute differentiation statistics \eqn{F_{ST}}, \eqn{G'_{ST}}, and \eqn{D}
for a list of data frames of allele frequencies or subclonal fractions.
The differentiation statistics are computed across all available subpopulations.
}
