% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diff_stats.R
\name{compute_mean_stats}
\alias{compute_mean_stats}
\title{Computes mean values and normalized mean valuesof FST, G'ST and D
statistics over all loci in a given sample.}
\usage{
compute_mean_stats(Diff_loci, K)
}
\arguments{
\item{Diff_loci}{A list of matrices containing values of differentiation statistics FST, G'ST, D
and a vector of frequencies M.}

\item{K}{Number of subpopulations}
}
\value{
A named list with six numeric values:
\describe{
\item{FST_mean}{Mean raw FST value across loci}
\item{GST_mean}{Mean raw G'ST value across loci}
\item{D_mean}{Mean raw Jost's D value across loci}
\item{FST_norm_mean}{Mean normalized FST value across loci}
\item{GST_norm_mean}{Mean normalized G'ST value across loci}
\item{D_norm_mean}{Mean normalized Jost's D value across loci}
}
}
\description{
Computes mean values and normalized mean valuesof FST, G'ST and D
statistics over all loci in a given sample.
}
\examples{
# Example with 2 loci and 2 subpopulations
library(tibble)
library(patchwork)
library(ggplot2)
library(glue)
locus1 <- matrix(c(0.8, 0.2,    # Subpop 1
                   0.6, 0.4),   # Subpop 2
                 nrow = 2, byrow = TRUE)

locus2 <- matrix(c(0.5, 0.5,
                   0.4, 0.6),
                 nrow = 2, byrow = TRUE)

list_loci <- list(locus1, locus2)
Diff_loci <- lapply(list_loci, Diff)
mean_stats <- compute_mean_stats(Diff_loci, K=2)
}
