% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGenBounds_functions.R
\name{Diff}
\alias{Diff}
\title{Computes differentiation statistics}
\usage{
Diff(freqs)
}
\arguments{
\item{freqs}{An allele frequency matrix, where rows are populations and columns alleles}
}
\value{
A numerical vector
}
\description{
Computes differentiation statistics
}
\examples{
freqs <- matrix(c(1,0.5,0,0.5),nrow=2)
Diff(freqs)
}
