% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGenBounds_functions.R
\name{Dup}
\alias{Dup}
\title{Computes the maximal D in terms of the most frequent allele}
\usage{
Dup(K, M)
}
\arguments{
\item{K}{Number of subpopulations}

\item{M}{Frequency of the most frequent allele}
}
\value{
A numerical value
}
\description{
Computes the maximal D in terms of the most frequent allele
}
\examples{
Dup(2,0.5)
}
