% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGenBounds_functions.R
\name{Fup}
\alias{Fup}
\title{Computes the maximal FST in terms of the most frequent allele}
\usage{
Fup(K, M)
}
\arguments{
\item{K}{Number of subpopulations}

\item{M}{Frequency of the most frequent allele}
}
\value{
A numerical value
}
\description{
Computes the maximal FST in terms of the most frequent allele
}
\examples{
#' Fup(2,0.5)
}
