% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diff_stats.R
\name{get_mean_stats}
\alias{get_mean_stats}
\title{Computes mean and normalized mean values of FST, G'ST, D for a list of samples.}
\usage{
get_mean_stats(data_frame_list, sample_names, K = 2)
}
\arguments{
\item{data_frame_list}{A list of dataframes containing allele fraquencies for loci (rows)
in multiple subpopulations (columns)}

\item{sample_names}{List of strings containing names of samples to include in the computation.}

\item{K}{Number of subpopulations}
}
\value{
A list of named lists containing six numeric values:
\describe{
\item{FST_mean}{Mean raw FST value across loci}
\item{GST_mean}{Mean raw G'ST value across loci}
\item{D_mean}{Mean raw Jost's D value across loci}
\item{FST_norm_mean}{Mean normalized FST value across loci}
\item{GST_norm_mean}{Mean normalized G'ST value across loci}
\item{D_norm_mean}{Mean normalized Jost's D value across loci}
} for a provided list of samples
}
\description{
Computes mean and normalized mean values of FST, G'ST, D for a list of samples.
}
