% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGenBounds_functions.R
\name{ggbounds}
\alias{ggbounds}
\title{Plots differentiation statistics along with the maximal and minimal values of the statistic}
\usage{
ggbounds(freqs)
}
\arguments{
\item{freqs}{An allele frequency matrix, where rows are populations and columns alleles}
}
\value{
A ggplot object
}
\description{
Plots differentiation statistics along with the maximal and minimal values of the statistic
}
\examples{
freqs <- matrix(c(1,0.5,0,0.5),nrow=2)
ggbounds(freqs)
}
