% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_bounds.R
\name{ggbounds_new}
\alias{ggbounds_new}
\title{Plot differentiation statistics values and their bounds (first column)
and the normalized values of statistics (second column)}
\usage{
ggbounds_new(M, FST, GpST = NULL, D = NULL, K = 2)
}
\arguments{
\item{M}{Numeric vector of \eqn{M} values, one per locus.}

\item{FST}{Numeric vector of \eqn{F_{ST}} values, one per locus.}

\item{GpST}{Optional numeric vector of \eqn{G'_{ST}} values.}

\item{D}{Optional numeric vector of Jost's \eqn{D} values.}

\item{K}{Number of populations.}
}
\value{
A list of ggplot objects
}
\description{
Plot differentiation statistics values and their bounds (first column)
and the normalized values of statistics (second column)
}
\examples{
library(tidyverse)
freqs_locus1 <- matrix(c(1,0.5,0,0.5),nrow=2)
freqs_locus2 <- matrix(c(1,0.8,0,0.2),nrow=2)
freqs_locus3 <- matrix(c(1,0.2,0,0.8),nrow=2)
data  <- rbind(Diff(freqs_locus1),Diff(freqs_locus2),Diff(freqs_locus2))
ggbounds_new(M=data \%>\% filter(statistic=="M") \%>\% pull(value),
             FST=data \%>\% filter(statistic=="FST") \%>\% pull(value))
}
