% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGenBounds_functions.R
\name{ggfreqtable}
\alias{ggfreqtable}
\title{Plots differentiation statistics along with the maximal and minimal values of the statistic}
\usage{
ggfreqtable(freqs)
}
\arguments{
\item{freqs}{a matrix of allele frequency}
}
\value{
A ggplot object
}
\description{
Plots differentiation statistics along with the maximal and minimal values of the statistic
}
\examples{
popa = matrix(c(0.3,0.3,0.3,0.1,rep(0,4*5),
  rep(0,4),0.3,0.3,0.3,0.1,rep(0,4*4),
  rep(0,4*2),0.3,0.3,0.3,0.1,rep(0,4*3),
  rep(0,4*3),0.3,0.3,0.3,0.1,rep(0,4*2),
  rep(0,4*4),0.3,0.3,0.3,0.1,rep(0,4),
  rep(0,4*5),0.3,0.3,0.3,0.1),ncol=24,byrow = TRUE)
ggfreqtable(popa)
}
