% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopGenBounds_functions.R
\name{Gpup}
\alias{Gpup}
\title{Computes the maximal G'ST in terms of the most frequent allele}
\usage{
Gpup(K, M)
}
\arguments{
\item{K}{Number of subpopulations}

\item{M}{Frequency of the most frequent allele}
}
\value{
A numerical value
}
\description{
Computes the maximal G'ST in terms of the most frequent allele
}
\examples{
Gpup(2,0.5)
}
