% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{make_figure_sup}
\alias{make_figure_sup}
\title{Function to produce a joint plot of differentiation statistics over samples.}
\usage{
make_figure_sup(diff_stats, samples_2pop, samples_3pop, is_normalised)
}
\arguments{
\item{diff_stats}{Named list of differentiation statistics values for each sample.}

\item{is_normalised}{Bool value indicating whether the statistics values were normalized by the corresponding upper bound}
}
\value{
ggplot object
}
\description{
Function to produce a joint plot of differentiation statistics over samples.
}
