% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2d_freq.R
\name{plot_2d_freq}
\alias{plot_2d_freq}
\title{Plot a 2d-scatter plot of allele frequencies for each pair of subpopulations in a sample}
\usage{
plot_2d_freq(df, K)
}
\arguments{
\item{df}{A dataframe with loci in rows and allele freqencies in subpopulations in columns}

\item{K}{Number of subpopulations}
}
\value{
A ggplot object
}
\description{
Plot a 2d-scatter plot of allele frequencies for each pair of subpopulations in a sample
}
