% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{plot_diff_over_samples}
\alias{plot_diff_over_samples}
\title{Make plots of distributions of differentiation statistics \eqn{F_{ST}}, \eqn{G'_{ST}}, \eqn{D}
and their normalized versions.}
\usage{
plot_diff_over_samples(diff_stats, K, is_normalised = FALSE)
}
\arguments{
\item{diff_stats}{Named list of differentiation statistics values for each sample.}

\item{K}{The number of subpopulations}

\item{is_normalised}{Boolean indicator whether to plot original statistics or their normalized versions.}
}
\value{
A named list of stacked plots, each element correspond to a stacked plot of 3 subplots,
corresponding for \eqn{F_{ST}}, \eqn{G'_{ST}}, \eqn{D} values plotted versus \eqn{M}.
}
\description{
Returns a list of plots of \eqn{F_{ST}}, \eqn{G'_{ST}}, \eqn{D} values versus
the mean frequency of the most frequent allele \eqn{M}.
}
