% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{plot_stats_3subpop}
\alias{plot_stats_3subpop}
\title{Plot differentiation statistics for a given sample in case of K = 3 subpopulations}
\usage{
plot_stats_3subpop(D_loci, subpop_names)
}
\arguments{
\item{D_loci}{Four lists of matrices containing differentiation statistics FST, G'ST, D and a vector of frequencies M.
The first list corresponds to 3 subpopulations together, the remaining lists
correspond to pair-wise differentiation statistics for each pair of subpopulations out of 3 subpopulations.}

\item{subpop_names}{A string vector containing names of subpopulations to be plotted.}
}
\value{
A stacked ggplot of four elements. The first plot corresponds to values of
differentiation statistics computed for \eqn{K=3} subpopulations, the next three plots
correspond to differentiation statistics computed across pairs of subpopulations.
}
\description{
Plot differentiation statistics for a given sample in case of K = 3 subpopulations
}
\examples{
library(patchwork)
library(glue)
library(tibble)
# Initialise allele frequencies for 3 subpopulations in 2 loci
locus1 <- matrix(c(0.8, 0.2,   # Subpop 1
                   0.6, 0.4,   # Subpop 2
                   0.7, 0.3),  # Subpop 3
                 nrow = 3, byrow = TRUE)

locus2 <- matrix(c(0.5, 0.5,
                   0.4, 0.6,
                   0.6, 0.4),
                 nrow = 3, byrow = TRUE)

# Compute differentiation statistics
diff_stats <- compute_Diff_3subpop(list(locus1, locus2))

# Produce plot of differentiation statistics values
final_plot <- plot_stats_3subpop(diff_stats, c("Subpop1", "Subpop2", "Subpop3"))
}
