% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_plots}
\alias{save_plots}
\title{Save to file plots of differentiation statistics or of allele frequency.}
\usage{
save_plots(subpop_names, combined_plot, type = "stats", K = 2)
}
\arguments{
\item{subpop_names}{A string vector of subpopulations names to plot.}

\item{combined_plot}{A ggplot object to plot.}

\item{type}{A type of the plot: differentiation statistics (type = "stats"),
or allele frequencies (type = "freqs")}

\item{K}{Number of subpopulations.}
}
\value{
None
}
\description{
Save to file plots of differentiation statistics or of allele frequency.
}
