% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Toad_microsat_freq}
\alias{Toad_microsat_freq}
\title{Yellow-bellied toad microsatellite data}
\format{
\subsection{\code{Toad_microsat_freq}}{

A list of 10 matrices of allele frequencies, each with 47 rows (subpopulations) and a number of columns (alleles):
\describe{
\item{row}{Subpopulation sampled}
\item{columm}{microsatellite allele}
}
}
}
\source{
<Prohl et al 2021>
}
\usage{
Toad_microsat_freq
}
\description{
Data from Prohl et al 2021
}
\keyword{datasets}
