% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unparse_columns}
\alias{unparse_columns}
\title{Transform a combined column in multiple separate columns}
\usage{
unparse_columns(data_list, idx = 3)
}
\arguments{
\item{data_list}{A list of dataframes}

\item{idx}{Index of the column to replace}
}
\value{
List of modified dataframes
}
\description{
If a dataframe column consist of two columns, replaces the combined column by the two separate columns.
}
