% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_subfraction_data.R
\name{filter_clonal}
\alias{filter_clonal}
\title{Filter data with subclonal frequencies to exclude clonal mutations}
\usage{
filter_clonal(df, K = 2, filter_fixed = T)
}
\arguments{
\item{df}{A dataframe with loci in rows and different metrics in columns.}

\item{K}{Number of subpopulations, K=2 by default.}

\item{filter_fixed}{bool value, defines whether to filter out loci with 0 frequencies in both subpopulations.}
}
\value{
A dataframe containing only rows with "Clonal" = FALSE.
}
\description{
Filter data with subclonal frequencies to exclude clonal mutations
}
\examples{
#' library(dplyr)
# Simulated example dataset
df <- tibble::tibble(
  chr = c("1", "1", "1"),
  pos = c(12345, 12346, 12347),
  sub.frac.Sample1 = c(0.2, 0, 0.5),
  sub.frac.Sample2 = c(0.3, 0, 0.5),
  Clonal = c("FALSE", "TRUE", "Uncertain")
)
filtered_data <- filter_clonal(df, K=2)
}
