% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_bounds.R
\name{ggbounds_norm}
\alias{ggbounds_norm}
\title{Plots differentiation statistics normalized by their maximal values}
\usage{
ggbounds_norm(M, FST, GpST = NULL, D = NULL, K = 2)
}
\arguments{
\item{M}{Numerical vector of frequency of the most frequent allele for each locus}

\item{FST}{Numerical vector of FST values for each locus}

\item{GpST}{Numerical vector of G prime ST values for each locus}

\item{D}{Numerical vector of D values for each locus}

\item{K}{Number of subpopulations}
}
\value{
A list of ggplot objects
}
\description{
Plots differentiation statistics normalized by their maximal values
}
\examples{
library(tidyverse)
freqs_locus1 <- matrix(c(1,0.5,0,0.5),nrow=2)
freqs_locus2 <- matrix(c(1,0.8,0,0.2),nrow=2)
freqs_locus3 <- matrix(c(1,0.2,0,0.8),nrow=2)
data  <- rbind(Diff(freqs_locus1),Diff(freqs_locus2),Diff(freqs_locus2))
ggbounds_norm(M=data \%>\% filter(statistic=="M") \%>\% pull(value),
              FST=data \%>\% filter(statistic=="FST") \%>\% pull(value))
}
