% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diff_stats.R
\name{plot_mean_stats_sep}
\alias{plot_mean_stats_sep}
\title{Plots mean values of FST, G'ST and D and their normalised mean values for a given list of samples.}
\usage{
plot_mean_stats_sep(mean_stats_list)
}
\arguments{
\item{mean_stats_list}{A list of named list containing six numeric values:
\describe{
\item{FST_mean}{Mean raw FST value across loci}
\item{GST_mean}{Mean raw G'ST value across loci}
\item{D_mean}{Mean raw Jost's D value across loci}
\item{FST_norm_mean}{Mean normalized FST value across loci}
\item{GST_norm_mean}{Mean normalized G'ST value across loci}
\item{D_norm_mean}{Mean normalized Jost's D value across loci}
}
Each named list corresponds to a different data sample.}
}
\value{
A ggplot object
}
\description{
Plots mean values of FST, G'ST and D and their normalised mean values for a given list of samples.
}
