% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2d_freq.R
\name{run_2d_freq_plotting}
\alias{run_2d_freq_plotting}
\title{Run a pipeline for plotting pair-wise allele frequencies of subpopulations for a list of samples}
\usage{
run_2d_freq_plotting(data_frame_list, K = 2)
}
\arguments{
\item{data_frame_list}{A list of dataframes containing allele fraquencies for loci (rows)
in multiple subpopulations (columns)}

\item{K}{Number of subpopulations}
}
\description{
Run a pipeline for plotting pair-wise allele frequencies of subpopulations for a list of samples
}
